==> 3.Merge ToposText and NER.ipynb <==
import pandas as pd
from tqdm import tqdm
import numpy as np

## open the file containing flairNER annotation in the NH
#flairNER_NH = pd.read_csv("/Users/u0154817/OneDrive - KU Leuven/Documents/KU Leuven/PhD project 'Greek Spaces in Roman Times'/Data_Extraction/Outputs/flair_NH.csv", delimiter = ',')
flairNER_NH = pd.read_csv("data/intermediate/flair_NH.csv", delimiter = ',')

flairNER_NH['topostext_id'] = '-'

flairNER_NH

## open the file containing ToposText annotation in the NH
ToposText_NH = pd.read_csv("data/intermediate/topostext_places_NH.csv", delimiter=",")

ToposText_NH

#this is horribly slow takes about 8 seconds per iteration: 
# do_print = False
# total_items = len(ToposText_NH['tagged_entity'])
# for i1, topostext in tqdm(enumerate(ToposText_NH['tagged_entity']), total=total_items):
    
#     if do_print:
#         print(i1,topostext)
    
#     reference_topostext = ToposText_NH['reference'][i1]
#     startpos_topostext = ToposText_NH['start_pos'][i1]
#     topostext_id = ToposText_NH['topostext_id'][i1]
        
#     for i2,flairNER in enumerate(flairNER_NH['token']):
        
#         if flairNER_NH['reference'][i2] == reference_topostext:
#             if flairNER_NH['start_pos'][i2] == startpos_topostext:
                
#                 if 'LOC' in flairNER_NH['flair_ner'][i2]: ## the entity was annotated by flairNER
#                     #flairNER_NH['topostext_id'][i2] = topostext_id ## add the ToposText ID results in chainedAssignmentError. patch below:
#                     flairNER_NH.loc[i2, 'topostext_id'] = topostext_id ## add the ToposText ID
#                     if do_print:
#                         print(i2, flairNER_NH['reference'][i2], flairNER_NH['start_pos'][i2], flairNER_NH['flair_ner'][i2], flairNER_NH['topostext_id'][i2])
                    
#                 else : 
                    
#                     topostext = str(topostext)
#                     topostext = topostext.split() ## split the ToposText annotation into tokens
                    
#                     if len(topostext) == 1: ## it consists of a single token
                        
#                         flairNER_NH.loc[i2, 'flair_ner'] = 'B-LOC'
#                         flairNER_NH.loc[i2,'topostext_id'] = topostext_id
#                         if do_print:
#                             print(i2, flairNER_NH['reference'][i2], flairNER_NH['start_pos'][i2], flairNER_NH['flair_ner'][i2], flairNER_NH['topostext_id'][i2])
                    
#                     else :
                        
#                         flag = False
                        
#                         for i3, token in enumerate(topostext):
                            
#                             if i3 > 0:
                                
#                                 if flairNER_NH['flair_ner'][i2+i3] == 'B-LOC':
                                    
#                                     flag = True ## the entity is annotated inside
#                                     flairNER_NH.loc[i2+i3, 'topostext_id'] = topostext_id
#                                     if do_print:
#                                         print(i2+i3, flairNER_NH['reference'][i2+i3], flairNER_NH['start_pos'][i2+i3], flairNER_NH['flair_ner'][i2+i3], flairNER_NH['topostext_id'][i2+i3])
                                    
#                         if flag == False:
                            
#                             flairNER_NH.loc[i2, 'flair_ner'] = 'B-LOC'
#                             flairNER_NH.loc[i2, 'topostext_id'] = topostext_id
#                             if do_print:
#                                 print(i2, flairNER_NH['reference'][i2], flairNER_NH['start_pos'][i2], flairNER_NH['flair_ner'][i2], flairNER_NH['topostext_id'][i2])

#refactored version of code above.
do_print = False

# Convert DataFrame columns to numpy arrays for faster access
reference_topostext_arr = np.array(ToposText_NH['reference'])
startpos_topostext_arr = np.array(ToposText_NH['start_pos'])
topostext_id_arr = np.array(ToposText_NH['topostext_id'])
tagged_entity_arr = np.array(ToposText_NH['tagged_entity'])

reference_flairNER_arr = np.array(flairNER_NH['reference'])
startpos_flairNER_arr = np.array(flairNER_NH['start_pos'])
flair_ner_arr = np.array(flairNER_NH['flair_ner'])
token_arr = np.array(flairNER_NH['token'])

# Ensure topostext_id column exists in flairNER_NH DataFrame
if 'topostext_id' not in flairNER_NH.columns:
    flairNER_NH['topostext_id'] = np.nan

topostext_id_flairNER_arr = np.array(flairNER_NH['topostext_id'])

total_items = len(tagged_entity_arr)
for i1, topostext in tqdm(enumerate(tagged_entity_arr), total=total_items):
    
    if do_print:
        print(i1, topostext)
    
    reference_topostext = reference_topostext_arr[i1]
    startpos_topostext = startpos_topostext_arr[i1]
    topostext_id = topostext_id_arr[i1]
    
    # Create a mask for rows where references and start positions match
    mask = (reference_flairNER_arr == reference_topostext) & (startpos_flairNER_arr == startpos_topostext)
    
    # Get the indices of matching rows
    matching_indices = np.where(mask)[0]
    
    for i2 in matching_indices:
        
        if 'LOC' in flair_ner_arr[i2]:  # The entity was annotated by flairNER
            flairNER_NH.loc[i2, 'topostext_id'] = topostext_id  # Add the ToposText ID
            if do_print:
                print(i2, reference_flairNER_arr[i2], startpos_flairNER_arr[i2], flair_ner_arr[i2], flairNER_NH.loc[i2, 'topostext_id'])
        
        else:
            topostext_tokens = str(topostext).split()  # Split the ToposText annotation into tokens
            
            if len(topostext_tokens) == 1:  # It consists of a single token
                flairNER_NH.loc[i2, 'flair_ner'] = 'B-LOC'
                flairNER_NH.loc[i2, 'topostext_id'] = topostext_id
                if do_print:
                    print(i2, reference_flairNER_arr[i2], startpos_flairNER_arr[i2], flair_ner_arr[i2], flairNER_NH.loc[i2, 'topostext_id'])
            else:
                flag = False
                
                for i3, token in enumerate(topostext_tokens):
                    if i3 > 0 and i2 + i3 < len(flair_ner_arr):
                        if flair_ner_arr[i2 + i3] == 'B-LOC':
                            flag = True  # The entity is annotated inside
                            flairNER_NH.loc[i2 + i3, 'topostext_id'] = topostext_id
                            if do_print:
                                print(i2 + i3, reference_flairNER_arr[i2 + i3], startpos_flairNER_arr[i2 + i3], flair_ner_arr[i2 + i3], flairNER_NH.loc[i2 + i3, 'topostext_id'])
                
                if not flag:
                    flairNER_NH.loc[i2, 'flair_ner'] = 'B-LOC'
                    flairNER_NH.loc[i2, 'topostext_id'] = topostext_id
                    if do_print:
                        print(i2, reference_flairNER_arr[i2], startpos_flairNER_arr[i2], flair_ner_arr[i2], flairNER_NH.loc[i2, 'topostext_id'])


flairNER_NH

flairNER_NH.drop_duplicates(keep='first', inplace=True, ignore_index=True)

flairNER_NH['reference'] = flairNER_NH['reference'].astype(str)
flairNER_NH.to_csv('data/intermediate/flairToposText_NH.csv', index=False)
