==> 5.Group_in_Sentences_xml_Clerice_to_csv.ipynb <==
Generate separated files for each book containing the list of sentences. Sentence divisions is performed in a semi-manual way using specific patterns: only full stop is included when it is preceded by a token longer than 2 digits (i.e., p., M., but also ea). Inconsistencies and errors are present in the sentence split. The sentence count is at a book level: this means that sentence 23 in Book 4 is the 23rd sentence in the book.

import pandas as pd
import re
import requests
from bs4 import BeautifulSoup

path = "data/intermediate/xml_Clerice_to_csv.csv"

xml_Clerice_to_csv = pd.read_csv(path, encoding='utf-8')

## extract the book from the reference and create a new column
book_column = []
pattern = r'(\d+)(?:\.(\d))?\.'

for i, reference in enumerate(xml_Clerice_to_csv['reference']):
    
    match = re.search(pattern,reference)
    book_column.append(int(match[1]))
    
xml_Clerice_to_csv['book'] = book_column

punctuation = ['.']

sentence_column = [] ## create a new column for the sentence count

for i in range(1,38): ## for each book of the NH
        
    # filter_book = LatLemm[LatLemm['book'] == i]
    filter_book = xml_Clerice_to_csv[xml_Clerice_to_csv['book'] == i]
    filter_book = filter_book.reset_index()
    
    count_sentence = 0
    
    for index,token in enumerate(filter_book['token']):
        sentence_column.append(count_sentence)
        
        if token in punctuation:
            
            if len(filter_book['token'][index-1]) < 3: ## if the full stop is preceded by a token shorter than 3 digits
                print(filter_book['index'][index-1], filter_book['token'][index-1]) ## do not increment the sentence count
                
            else : count_sentence = count_sentence + 1

xml_Clerice_to_csv['sentence'] = sentence_column

xml_Clerice_to_csv.to_csv('data/intermediate/NH_lat_groupedsentences.csv', index=False)

## test
## filter all the rows containing sentence = 0
filter_book37 = xml_Clerice_to_csv[xml_Clerice_to_csv['book'] == 37]
print_sent_0 = filter_book37[filter_book37['sentence'] == 0]

## concatenate the tokens in filtered_rows
concatenate_string = ' '.join(print_sent_0['token'].astype(str))
concatenate_string

## write files containing the sentences
for i in range(1,38): ## for each book of the NH
        
    filter_book = xml_Clerice_to_csv[xml_Clerice_to_csv['book'] == i]
    name='data/intermediate/books/latbook'+str(i)
    with open(name, "w+", encoding='utf-8') as file:
        for i in range(0, filter_book['sentence'].max()+1):
            filter_rows = filter_book[filter_book['sentence'] == i]
            concatenate_sentence = ' '.join(filter_rows['token'].astype(str))
            file.write(f"{concatenate_sentence}\n")

# print('NOW YOU NEED TO RUN Use_SimAlign_Tokens_and_Project_annotations.ipynb OR vecalign.py?? ')

#takes input generated in the Use_SimAlign_Tokens_and_Project_annotations.ipynb notebook. 
# path = "data/intermediate/NH_lat_projected.csv"

# NH_projected = pd.read_csv(path, encoding='utf-8', dtype={'reference': 'str'})

# NH_projected

# index_in_sentence_column = []
# index_in_sentence = 0

# for i,token in enumerate(NH_projected['token']):
#     print(i)
#     if i == 0:
#         index_in_sentence_column.append(index_in_sentence)
#     if i > 0:
#         if NH_projected['sentence'][i] == NH_projected['sentence'][i-1]: ## it is still in the same sentence
#             index_in_sentence = index_in_sentence+1 ## increment
#             index_in_sentence_column.append(index_in_sentence)
#         else : 
#             index_in_sentence = 0 # reset the index
#             index_in_sentence_column.append(index_in_sentence)

# len(index_in_sentence_column)

# NH_projected['index_in_sentence'] = index_in_sentence_column

# NH_projected

# NH_projected.to_csv('NH_projected_index_sentences.csv',index=False)

# NH_projected_sentences = NH_projected[['index_in_sentence']].copy()

# NH_projected_sentences

# NH_projected_sentences.to_csv('index_in_sentences.csv',index=False)

# path = "/Users/u0154817/OneDrive - KU Leuven/Documents/KU Leuven/PhD project 'Greek Spaces in Roman Times'/Data_Extraction/Outputs/table_book_(chapter)_chapter_paragraph.csv"

# table_of_references = pd.read_csv(path, delimiter=';', encoding='utf-8', dtype={'reference': 'str'})

# table_of_references.dtypes

# table_of_references

# book_column = []
# chapter_column = []
# chapter_in_bracket_column = []

# for book in range(1,38): ## for each book of the NH
        
#     filter_book = table_of_references[table_of_references['Book'] == book]
#     filter_book = filter_book.reset_index()
    
#     unique_chapters = filter_book['Chapter'].unique().tolist() ## get a unique list of chapter numbers
#     chapter_column.extend(unique_chapters)
#     print(book, unique_chapters)
    
#     for chapter in unique_chapters:
#         book_column.append(book)
#         temp_chapter_in_brackets = []
        
#         for i,value in enumerate(filter_book['Chapter']):
#             if value == chapter:
#                 if filter_book['(Chapter)'][i] in temp_chapter_in_brackets:
#                     continue
#                 else : temp_chapter_in_brackets.append(filter_book['(Chapter)'][i])
        
#         chapter_in_bracket_column.append(temp_chapter_in_brackets)       

# reference = pd.DataFrame({'book': book_column, 'chapter_Perseus': chapter_column, '(chapters)_ToposText': chapter_in_bracket_column})
# reference.head(50)

# reference.to_csv('chapter_correspondences.csv',index=False)
