==> 6.Group_in_Sentences_Eng.ipynb <==
Tokens are grouped in enumerated sentences in order to perform sentence alignment. For each book, sentences are saved in a separated file. Moreover, the code assigns a number (_loc_ent_id_) to each LOC entity (B-LOC and I-LOC) in the NH. In the case the loc_ent_id is associated with a ToposText ID, a query is made to get also the TrismegistosPlace ID.

import pandas as pd
import re
from tqdm import tqdm

path = "data/intermediate/flairToposText_NH.csv"

## open the file containing flairNER annotation in the NH
flairNER_NH = pd.read_csv(path, delimiter = ',')

flairNER_NH.head(60)

book_column = [] ## extract the book reference
pattern = r'(\d+)(?:\.(\d))?\.'

for i, reference in enumerate(flairNER_NH['reference']):
    try  : 
        match = re.search(pattern,reference)
        book = int(match[1])
    except : book = 11 
    book_column.append(book)
        
flairNER_NH['book'] = book_column

flairNER_NH['book'].unique()

## exclude book 0 which contains a preface to the NH
flairNER_NH = flairNER_NH.loc[~(flairNER_NH ['book'] == 0)]
flairNER_NH.reset_index(inplace=True)
flairNER_NH = flairNER_NH.drop('level_0', axis=1)

flairNER_NH[flairNER_NH.isnull().any(axis=1)]
## we observed that 'None' tokens were interpreted as None elements
flairNER_NH['token'] = flairNER_NH['token'].fillna('None')
#flairNER_NH.token.fillna('None', inplace=True) 

punctuations = ['.', '!', '?'] ## set the punctuation marks

Starting from sentence number = 0, each token is assigned to the sentence number inside the book. If the token is a full stop, an exclamation mark or a question mark, the following token has sentence number = sentence number + 1.

sentence_column = []

for i1 in range(1,38): ## for each book of the NH
        
    filter_book = flairNER_NH[flairNER_NH['book'] == i1] ## select all the tokens in the book
    filter_book = filter_book.reset_index()
    
    count_sentence = 0 ## set the sentence number to 0
    
    for i2,token in enumerate(filter_book['token']):
        sentence_column.append(count_sentence) ## assign the token to the sentence number     
        if token in punctuations: ## if the token is a punctuation sign increment the sentence number
            count_sentence = count_sentence + 1

flairNER_NH['sentence'] = sentence_column

flairNER_NH.to_csv(r"data/intermediate/NH_eng_groupedsentences.csv", index=False)

## test
## filter all the rows containing sentence = 0 in book 1
filter_book1 = flairNER_NH[flairNER_NH['book'] == 1]
print_sent_0 = filter_book1[filter_book1['sentence'] == 0]

## concatenate the tokens in print_sent_0
concatenate_string = ' '.join(print_sent_0['token'].astype(str))
concatenate_string

## write files containing the sentences
for i in range(1,38): ## for each book of the NH
        
    filter_book = flairNER_NH[flairNER_NH['book'] == i] ## select all the tokens in the book
    name='data/intermediate/books/engbook'+str(i)
    with open(name, "w", encoding='utf-8') as file: ## create a file for the book
        for i in range(0, filter_book['sentence'].max()+1): ## for each sentence in the book
            filter_rows = filter_book[filter_book['sentence'] == i] ## select all the tokens in the sentence
            concatenate_sentence = ' '.join(filter_rows['token'].astype(str)) ## concatenate the tokens in a sentence
            file.write(f"{concatenate_sentence}\n") ## write the sentence in the file

flairNER_NH['loc_ent_id'] = '-' ## create a new column for the id of the LOC entity

count_loc = -1
for i,loc in enumerate(flairNER_NH['flair_ner']):
    if 'LOC' in loc: ## for each LOC entity
        if loc == 'B-LOC': ## if it is a B-LOC entity
            count_loc = count_loc+1 ## increment + 1
            flairNER_NH.loc[i,'loc_ent_id'] = count_loc ## assign the LOC id
        else : flairNER_NH.loc[i, 'loc_ent_id'] = count_loc ## assign the LOC id

loc_ent_ids = [element for element in flairNER_NH['loc_ent_id'].unique() if isinstance(element, int)]

loc_ent_column = []
topostext_id = []

for n in tqdm(range(min(loc_ent_ids), max(loc_ent_ids)+1)):
    loc_ent_column.append(n)
    #print(n)
    filter_loc_entity = flairNER_NH[flairNER_NH['loc_ent_id'] == n] ## select all the tokens in the entity
    filter_loc_entity = filter_loc_entity.reset_index()
    for i,token in enumerate(filter_loc_entity['token']):
        
        flag = False
        
        if filter_loc_entity['topostext_id'][i] != '-':
            
            topostext_id.append(filter_loc_entity['topostext_id'][i])
            #print(n, filter_loc_entity['topostext_id'][i])
            flag = True
            break
            
    if flag == False:
        topostext_id.append('-')
        #print(n, filter_loc_entity['topostext_id'][i])

loc_ent_id = pd.DataFrame(
{'loc_ent_id': loc_ent_column,
'topostext_id': topostext_id
})

loc_ent_id.to_csv('loc_ent_ids.csv', index=False)

import requests
from bs4 import BeautifulSoup

#TODO move to a later phase. 
## import the trismegistosplace id 
topostext_id_column=[]
trismegistos_id_column=[]

requestamount = len(list(enumerate(loc_ent_id.topostext_id.unique())))
for i, topostext_id in tqdm(enumerate(loc_ent_id.topostext_id.unique()), total = requestamount): ## for each topostext_id
    #print(i)
    if topostext_id !='-': ## excluding the case of missing topostext_id
        
        topostext_id_column.append(topostext_id)
        temporary_list_TM=[]
        
        ## navigate the ToposText page of the place name
        topostext_page_name = 'https://topostext.org/place/'+str(topostext_id)
        topostext_page=requests.get(topostext_page_name)
        soup=BeautifulSoup(topostext_page.content)
        
        for a_tag in soup.find_all('a'):
            if "Trismegistos" in a_tag.contents[0]: ## if the TM id is present in the page
                trismegistos_id=a_tag.contents[0]
                temporary_list_TM.append(trismegistos_id)
    
        trismegistos_id_column.append(temporary_list_TM)

topostext_trismegistos = pd.DataFrame(
{'topostext_id': topostext_id_column,
'trismegistos_id': trismegistos_id_column
})

topostext_trismegistos

trismegistos_id_column = []
pattern = r': (.*)$'

for i, trismegistos_id in enumerate(topostext_trismegistos['trismegistos_id']):
    if len(trismegistos_id) == 1:
        for TM_id in trismegistos_id:
            match = re.search(pattern, TM_id)
            TM_id = match.group(1)
            print(TM_id)
    else: TM_id = '-'
        
    trismegistos_id_column.append(TM_id)

topostext_trismegistos['trismegistos_id'] = trismegistos_id_column

topostext_trismegistos

#topostext_trismegistos = topostext_trismegistos.drop('topostext_trismegistos', axis=1)

topostext_trismegistos.to_csv('data/intermediate/topostext_trismegistos_ids.csv', index=False)
